+++ services/api/src/main/java/com/stacta/api/fragrance/dto/FragranceRatingSummary.java
@@ -3,5 +3,5 @@ package com.stacta.api.fragrance.dto;
 public record FragranceRatingSummary(
   double average,
   long count,
-  Integer userRating
+  Double userRating
 ) {}
