+++ services/api/src/main/java/com/stacta/api/fragrance/dto/FragranceSearchResult.java
@@ -42,5 +42,5 @@ public record FragranceSearchResult(
   UUID createdByUserId,
   String createdByUsername,
   Long ratingCount,
-  Integer userRating
+  Double userRating
 ) {}
